+++ R/zzz.R
@@ -14,7 +14,7 @@ img_read_memoised <- img_read
 
 #' @import grid
 #' @importFrom glue glue
-#' @importFrom grDevices col2rgb dev.off png rgb
+#' @importFrom grDevices col2rgb dev.capabilities dev.off png rgb
 #' @importFrom rlang %||% abort inform warn
 #' @importFrom utils hasName head packageVersion tail
 NULL